+++ NEWS.md
@@ -3,6 +3,7 @@
 * Adding "Suggests" pkgs to Dependencies tab (#624)
 * Migrated from `{jsTreeR}` to `{shinyTree}` (#585)
 * Fix bug with the logo's hyperlink not functioning in the expected zone (#633)
+* Summarize Package Dependencies (#618)
 
 # riskassessment 2.0.0